+++ web/app/scripts/src/directives/form.es.js
@@ -1,5 +1,6 @@
 import { angular } from '../libraries';
 
+import API from '../services/api';
 import FieldsetParser from '../services/fieldsets-parser';
 
 export const moduleProperties = {
@@ -7,25 +8,54 @@ export const moduleProperties = {
 };
 
 const module = angular.module(moduleProperties.moduleName, [
+    API.moduleName,
     FieldsetParser.moduleName
 ]);
 
 module.directive('surveysForm', [
+    API.componentName,
     FieldsetParser.componentName,
-    function (FieldsetParserService) {
+    function (
+        API,
+        FieldsetParser
+    ) {
         return {
             restrict: 'A',
             scope: {
-                formStructure: '='
+                formUrl: '=',
+                responseUrl: '='
             },
-            template: '<formly-form model="model" fields="fields"></formly-form>',
-            link: function ($scope, $element, $attrs) {
-                $scope.$watch('formStructure', (form) => {
-                    if (form) {
-                        $scope.fields = FieldsetParserService.parseFields(form);
-                        $scope.model = FieldsetParserService.parseModel(form);
+            templateUrl: 'templates/incuna-surveys/survey-form.html',
+            link: function ($scope) {
+                $scope.$watch('formUrl', (url) => {
+                    if (url) {
+                        API.getForm(url).then(function (structure) {
+                            $scope.fields = FieldsetParser.parseFields(structure);
+                            // Only set the empty model if the model has not
+                            // been set.
+                            if (Object.keys($scope.model).length === 0) {
+                                $scope.model = FieldsetParser.parseFormToModel(structure);
+                            }
+                        });
                     }
                 });
+
+                $scope.$watch('responseUrl', (url) => {
+                    if (url) {
+                        API.get(url).then(function (data) {
+                            $scope.model = FieldsetParser.parseResponseToModel(data);
+                        });
+                    }
+                });
+
+                $scope.submit = () => {
+                    if ($scope.responseUrl) {
+                        // TODO: Handle errors.
+                        API.post($scope.responseUrl, $scope.model);
+                    }
+                }
+
+                $scope.submit();
             }
         };
     }