+++ web/tests/api-spec.js
@@ -11,11 +11,20 @@ describe('api service', function () {
             this.api = API;
         });
 
+        this.post = fixture.load('forms/pk/respond/post.json')
+        this.get = fixture.load('forms/pk/respond/user_id/get.json')
+
         this.$httpBackend.when('GET', 'localhost:8000/forms')
             .respond(fixture.load('forms/get.json').OK.response_data);
 
         this.$httpBackend.when('GET', 'http://localhost:8000/forms/1')
             .respond(fixture.load('forms/pk/get.json').OK.response_data);
+
+        this.$httpBackend.when('GET', this.get.url)
+            .respond(this.get.OK.response_data);
+        
+        this.$httpBackend.when('POST', this.post.url)
+            .respond(this.post.OK.response_data);
         
     });
 
@@ -52,6 +61,28 @@ describe('api service', function () {
             this.$httpBackend.flush();
         });
         
+    });
+
+    describe('get', function () {
+        it('should return a promise with the data from the api', function () {
+            this.api.getForm(this.get.url).then((data) => {
+                expect(data).toEqual(this.get.OK.response_data);
+            });
+
+            this.$rootScope.$digest();
+            this.$httpBackend.flush();
+        });
+    });
+
+    describe('post', function () {
+        it('should return a promise with the data of a new response', function () {
+            this.api.post(this.post.url, this.post.fields).then((data) => {
+                expect(data).toEqual(this.post.OK.response_data);
+            });
+
+            this.$rootScope.$digest();
+            this.$httpBackend.flush();
+        });
     });
     
 });